requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aioredis==1.3.1
    # via channels-redis
amqp==5.1.1
    # via kombu
anyio==3.6.1
    # via watchgod
asgiref==3.5.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   uvicorn
asttokens==2.0.5
    # via stack-data
async-timeout==4.0.2
    # via aioredis
attrs==21.4.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==22.4.2
    # via daphne
automat==20.2.0
    # via twisted
backcall==0.2.0
    # via ipython
billiard==3.6.4.0
    # via celery
celery==5.2.6
    # via
    #   -r requirements.in
    #   django-celery-beat
cffi==1.15.0
    # via cryptography
channels==3.0.4
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==3.4.0
    # via -r requirements.in
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
constantly==15.1.0
    # via twisted
cryptography==37.0.2
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==3.0.2
    # via channels
decorator==5.1.1
    # via ipython
django==3.2.13
    # via
    #   -r requirements.in
    #   channels
    #   django-celery-beat
    #   django-cors-headers
    #   django-extensions
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
django-celery-beat==2.2.1
    # via -r requirements.in
django-cors-headers==3.12.0
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-timezone-field==4.2.3
    # via django-celery-beat
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-extensions
djangorestframework-simplejwt==5.1.0
    # via -r requirements.in
drf-extensions==0.7.1
    # via -r requirements.in
executing==0.8.3
    # via stack-data
gunicorn==20.1.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
hiredis==2.0.0
    # via aioredis
httptools==0.4.0
    # via uvicorn
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.3
    # via
    #   anyio
    #   hyperlink
    #   twisted
incremental==21.3.0
    # via twisted
ipython==8.3.0
    # via -r requirements.in
jedi==0.18.1
    # via ipython
kombu==5.2.4
    # via celery
matplotlib-inline==0.1.3
    # via ipython
msgpack==1.0.3
    # via channels-redis
parso==0.8.3
    # via jedi
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.6.1
    # via -r requirements.in
prompt-toolkit==3.0.29
    # via
    #   click-repl
    #   ipython
psycopg2==2.9.3
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.21
    # via cffi
pygments==2.12.0
    # via ipython
pyjwt==2.4.0
    # via djangorestframework-simplejwt
pyopenssl==22.0.0
    # via twisted
python-crontab==2.6.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via python-crontab
python-dotenv==0.20.0
    # via
    #   -r requirements.in
    #   uvicorn
pytz==2022.1
    # via
    #   celery
    #   django
    #   django-timezone-field
    #   djangorestframework
pyyaml==6.0
    # via uvicorn
service-identity==21.1.0
    # via twisted
six==1.16.0
    # via
    #   asttokens
    #   automat
    #   click-repl
    #   python-dateutil
    #   service-identity
sniffio==1.2.0
    # via anyio
sqlparse==0.4.2
    # via django
stack-data==0.2.0
    # via ipython
tomli==2.0.1
    # via pep517
traitlets==5.2.0
    # via
    #   ipython
    #   matplotlib-inline
twisted[tls]==22.4.0
    # via daphne
txaio==22.2.1
    # via autobahn
typing-extensions==4.2.0
    # via twisted
uvicorn[standard]==0.17.6
    # via -r requirements.in
uvloop==0.16.0
    # via uvicorn
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchgod==0.8.2
    # via uvicorn
wcwidth==0.2.5
    # via prompt-toolkit
websockets==10.3
    # via uvicorn
wheel==0.37.1
    # via pip-tools
zope-interface==5.4.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
